requirements.txt
autocommand==2.2.2
backports.tarfile==1.2.0
cheroot==10.0.1
CherryPy==18.10.0
jaraco.collections==5.2.1
jaraco.context==6.0.1
jaraco.functools==4.2.1
jaraco.text==4.0.0
more-itertools==10.7.0
numpy==2.3.2
pandas==2.3.1
portend==3.2.1
pyspellchecker==0.8.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tempora==5.8.1
tzdata==2025.2
zc.lockfile==3.0.post1
